+++ packages/shared/src/constants/http-headers.ts
@@ -1,0 +1,5 @@
+export const MSL_TRAIN_ID_HEADER = 'MSL-Train-Id'
+export const MSL_TRAIN_ID_HEADER_LOWER = MSL_TRAIN_ID_HEADER.toLowerCase()
+
+export const MSL_ACCOUNT_HEADER = 'MSL-Account'
+export const MSL_ACCOUNT_HEADER_LOWER = MSL_ACCOUNT_HEADER.toLowerCase()